グラフを描画する箇所を外部ユーティリティに切り分け
